Split Process and InfraFunction into own files
refactor(function): split data/input/process.py into
 data/input/function/process.py and
 data/input/function/infrafunction.py;
 [Process] keep ingress/egress/integration_cache and the pure
 process_0/process_1 transfer functions, now returning the Ray
 Dataset directly with no MinIO/output-path knowledge;
 [InfraFunction] move infrafunction_subproc, its Ray Job dispatch,
 and the entrypoint script delivering Process's Dataset into
 MinIO into their own actuator module;
 [Structure] update infrastructure module's main.tf comment for
 the new infrafunction.py path;
 [Node] update sys.path comment for the split module paths;
 update cats_demo.py, tests/verification_test.py, docs/BOM.md,
 and .gitkeep for the new data/input/function/ layout
