+++ docs/BOM.md
@@ -27,7 +27,7 @@ The resulting BOM then pairs each of those *specified-as-code* CIDs with an *obs
   {"minio_bucket":"cats-scratch","minio_endpoint_host":"http://127.0.0.1:9000","minio_endpoint_pod":"http://172.19.0.1:9000"}
   ```
 
-Concretely, `InfraFunction` dispatches `integrated_subproc` as a real Ray Job (via the Ray Job Submission API) onto the Plant's live Ray cluster rather than running it in-process, and that job writes its output blocks directly to the shared MinIO bucket from whichever node executes each write task — genuinely distributed, rather than gathering the whole result set onto one node first (`data/input/process.py`'s `infrafunction_subproc`/`_run_ray_batches`). `bom.infrastructure_snapshot_cid` is what lets a downstream verifier confirm which shared store that distributed write actually landed in.
+Concretely, `InfraFunction` (`data/input/function/infrafunction.py`'s `infrafunction_subproc`, the actuator) dispatches `integrated_subproc` (Process's `_run_ray_batches`, a pure transfer function in `data/input/function/process.py` with no knowledge of MinIO) as a real Ray Job via the Ray Job Submission API onto the Plant's live Ray cluster rather than running it in-process; the job's entrypoint script (`infrafunction.py`'s `_INFRAFUNCTION_ENTRYPOINT_SCRIPT`) then delivers the returned Dataset by writing its output blocks directly to the shared MinIO bucket from whichever node executes each write task — genuinely distributed, rather than gathering the whole result set onto one node first. `bom.infrastructure_snapshot_cid` is what lets a downstream verifier confirm which shared store that distributed write actually landed in.
 
 Neither snapshot CID is re-consumed downstream to drive further behavior — their purpose is purely to make the executed Plant/InfraStructure state part of the CAT's permanent, content-addressed record. `flatten_bom()` currently only fetches `plant_snapshot_cid` back out into `flat_bom['plant']` for human-readable inspection; `infrastructure_snapshot_cid` isn't flattened the same way yet, so it's only reachable by `ipfs cat`-ing it directly out of the raw `bom` dict.
